Moved WindowForWrapper into WindowWrapper.cpp and moved header for WindowForWrapper into Window.h
